+++ tools/rosetta/client_online.go
@@ -22,8 +22,8 @@ import (
 	"github.com/tendermint/tendermint/rpc/client/http"
 	"google.golang.org/grpc"
 
-	crgerrs "github.com/cosmos/cosmos-sdk/server/rosetta/lib/errors"
-	crgtypes "github.com/cosmos/cosmos-sdk/server/rosetta/lib/types"
+	crgerrs "cosmossdk.io/tools/rosetta/lib/errors"
+	crgtypes "cosmossdk.io/tools/rosetta/lib/types"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	grpctypes "github.com/cosmos/cosmos-sdk/types/grpc"
@@ -31,6 +31,7 @@ import (
 	auth "github.com/cosmos/cosmos-sdk/x/auth/types"
 	bank "github.com/cosmos/cosmos-sdk/x/bank/types"
 
+	"github.com/cosmos/cosmos-sdk/types/query"
 	tmrpc "github.com/tendermint/tendermint/rpc/client"
 )
 
@@ -131,9 +132,7 @@ func (c *Client) Ready() error {
 		return err
 	}
 
-	// to prevent timeout of reading genesis block
-	var height int64 = -1
-	_, err = c.BlockByHeight(ctx, &height)
+	_, err = c.tmRPC.Status(ctx)
 	if err != nil {
 		return err
 	}
@@ -202,12 +201,13 @@ func (c *Client) BlockByHash(ctx context.Context, hash string) (crgtypes.BlockRe
 
 func (c *Client) BlockByHeight(ctx context.Context, height *int64) (crgtypes.BlockResponse, error) {
 	height, err := c.getHeight(ctx, height)
+
 	if err != nil {
 		return crgtypes.BlockResponse{}, crgerrs.WrapError(crgerrs.ErrBadGateway, err.Error())
 	}
 	block, err := c.tmRPC.Block(ctx, height)
 	if err != nil {
-		return crgtypes.BlockResponse{}, crgerrs.WrapError(crgerrs.ErrBadGateway, err.Error())
+		return crgtypes.BlockResponse{}, crgerrs.WrapError(crgerrs.ErrInternal, err.Error())
 	}
 
 	return c.converter.ToRosetta().BlockResponse(block), nil
@@ -237,11 +237,31 @@ func (c *Client) BlockTransactionsByHeight(ctx context.Context, height *int64) (
 
 // Coins fetches the existing coins in the application
 func (c *Client) coins(ctx context.Context) (sdk.Coins, error) {
+	var result sdk.Coins
+
 	supply, err := c.bank.TotalSupply(ctx, &bank.QueryTotalSupplyRequest{})
 	if err != nil {
 		return nil, crgerrs.FromGRPCToRosettaError(err)
 	}
-	return supply.Supply, nil
+
+	pages := supply.GetPagination().GetTotal()
+	for i := uint64(0); i < pages; i++ {
+		// get next key
+		page := supply.GetPagination()
+		if page == nil {
+			return nil, crgerrs.WrapError(crgerrs.ErrCodec, fmt.Sprintf("error pagination"))
+		}
+		nextKey := page.GetNextKey()
+
+		supply, err = c.bank.TotalSupply(ctx, &bank.QueryTotalSupplyRequest{Pagination: &query.PageRequest{Key: nextKey}})
+		if err != nil {
+			return nil, crgerrs.FromGRPCToRosettaError(err)
+		}
+
+		result = append(result[:0], supply.Supply[:]...)
+	}
+
+	return result, nil
 }
 
 func (c *Client) TxOperationsAndSignersAccountIdentifiers(signed bool, txBytes []byte) (ops []*rosettatypes.Operation, signers []*rosettatypes.AccountIdentifier, err error) {